Move web app client files into src/client
