+++ app/api/.gitignore
@@ -1,0 +1,1 @@
+src/app_api_debug
