+++ app/webapp/src/client/App.tsx
@@ -1,0 +1,28 @@
+import React from 'react';
+import { RESTAPI } from "./components/Sample";
+import logo from './logo.svg';
+import './App.css';
+
+const App: React.FC = () => {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+          Edit <code>src/App.tsx</code> and save to reload.1
+        </p>
+        <a
+          className="App-link"
+          href="https://reactjs.org"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Learn React
+        </a>
+        <RESTAPI></RESTAPI>
+      </header>
+    </div>
+  );
+}
+
+export default App;
