+++ app/webapp/src/server/index.ts
@@ -1,0 +1,2 @@
+import React from 'react';
+import ReactDOMServer from 'react-dom/server';
