+++ app/webapp/src/server/template.ts
@@ -1,0 +1,14 @@
+export default (html) => `
+    <!DOCTYPE html>
+    <html>
+    <head>
+        <title>test app</title>
+        <meta charset="utf-8">
+        <meta name="viewport" content="width=device-width,initial-scale=1,maximum-scale=1,user-scalable=no" />
+    </head>
+    <body>
+        <div id="app">${html}</div>
+        <script src="/bundle.min.js"></script>
+    </body>
+    </html>
+`;
